Centralize build
This adds a build.ps1 that builds all projects in the repo. It also adds props
file that centralizes the version number of the ML.NET package that should be
used. It currently also adds a nuget.config that inlcudes our nightlies so
that we can test against the latest API in case we have to.